Update utils.py
I'm not sure why you put that 1033 inside that 1036 check that returns false if the file is not saved (""). So it was preventing the automatic save from working correctly if the blend file was not saved. This is my first contrib and I don't understand the code very well, so sorry if it's bullshit :).